harvester-main/requirements.txt
couchdb
Mastodon.py
mpi4py
textblob